Move copy url button to its own component
